Migrate bindSocket into net-utils
