+++ .changeset/bright-buttons-smile.md
@@ -1,0 +1,7 @@
+---
+'@arcanewizards/net-utils': patch
+---
+
+Introduce bindSocket utility function
+
+Migrate the `bindSocket` utility function from the artnet package.
